# openapi.yaml
openapi: 3.1.0
info:
  title: classroomiq — API
  version: "0.0.1"
  license:
    name: Propietario
  description: |
    Contrato de la API del backend de **classroomiq**, plataforma de asistencia a evaluación docente
    universitaria. Fuente de verdad para el frontend.

    **Principio inamovible:** el docente es el juez; el sistema nunca asigna la nota final, genera
    borradores que el docente revisa, ajusta y aprueba.

    ## Autenticación
    Todas las rutas excepto `POST /api/auth/login` y `/actuator/health` exigen un **JWT Bearer**
    obtenido en el login. El token porta `sub` (usuarioId), `tenant_id`, `rol` y `email`. La
    autorización es por rol (`ADMIN`, `DOCENTE`, `COORDINADOR`).

    ## Aislamiento
    Multi-tenant: cada institución (tenant) y cada docente operan aislados. Un recurso de otro
    tenant/docente responde `404` (no se distingue de "no existe", para no filtrar existencia).

    ## Errores
    Respuestas de error en formato **RFC 7807 `application/problem+json`** (`ProblemDetail`). En
    validación de cuerpo (`400`) se agregan propiedades por campo con el mensaje de error.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/lotes/{loteId}/entregas:
    post:
      operationId: postLotesByLoteIdEntregas
      tags: [Lotes y entregas]
      summary: Subir una entrega (multipart)
      description: |
        Sube los archivos de un estudiante. La clasificación valida coherencia con `tipo`
        (pdf/docx → DOCUMENTO, zip → CODIGO, ambos → MIXTA). La entrega queda en estado `PENDIENTE`.
      parameters: [{ $ref: "#/components/parameters/LoteId" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [identificadorEstudiante, tipo, archivos]
              properties:
                identificadorEstudiante:
                  type: string
                  description: Identificador o alias del estudiante/grupo (no requiere datos personales).
                tipo: { $ref: "#/components/schemas/TipoEntrega" }
                archivos:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "201":
          description: Entrega creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntregaResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    get:
      operationId: getLotesByLoteIdEntregas
      tags: [Lotes y entregas]
      summary: Listar entregas de un lote
      parameters: [{ $ref: "#/components/parameters/LoteId" }]
      responses:
        "200":
          description: Entregas.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EntregaResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/entregas/{id}:
    get:
      operationId: getEntregasById
      tags: [Lotes y entregas]
      summary: Obtener entrega
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Entrega.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntregaResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteEntregasById
      tags: [Lotes y entregas]
      summary: Eliminar entrega
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Eliminada. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/entregas/{id}/contenido:
    get:
      operationId: getEntregasByIdContenido
      tags: [Lotes y entregas]
      summary: Contenido completo de la entrega
      description: |
        Texto íntegro de la entrega, reconstruido bajo demanda (re-extracción de los archivos en
        disco; sin costo de LLM/embeddings). Lo usa el panel de revisión para mostrar el documento
        completo con las citas del motor resaltadas en su contexto.
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Contenido por archivo y sección.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContenidoEntregaResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    LoteId:
      name: loteId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Falta el token o es inválido.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Forbidden:
      description: El rol no tiene permiso para la operación.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    NotFound:
      description: Recurso inexistente o de otro tenant/docente.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
    UnprocessableEntity:
      description: Regla de negocio violada (datos coherentes en forma pero inválidos en dominio).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    PayloadTooLarge:
      description: La subida supera el límite configurado.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
  schemas:
...
    TipoEntrega:
      type: string
      enum: [DOCUMENTO, CODIGO, MIXTA]
    EstadoEntrega:
      type: string
      enum: [PENDIENTE, PROCESANDO, EVALUANDO, LISTO, ERROR]
...
    RolArchivo:
      type: string
      enum: [DOCUMENTO, CODIGO]
...
    ProblemDetail:
      type: object
      description: |
        RFC 7807. En errores de validación (`400`) se agregan propiedades dinámicas, una por campo
        inválido, con el mensaje como valor.
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
      additionalProperties: true

    # ---- Auth / Usuario
...
    EntregaResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        loteId: { type: string, format: uuid }
        identificadorEstudiante: { type: string }
        tipo: { $ref: "#/components/schemas/TipoEntrega" }
        estado: { $ref: "#/components/schemas/EstadoEntrega" }
        mensajeError: { type: [string, "null"], description: "Descripción del fallo si estado = ERROR." }
        archivos:
          type: array
          items: { $ref: "#/components/schemas/ArchivoResponse" }
    ArchivoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombreOriginal: { type: string }
        mimeType: { type: string }
        tamanoBytes: { type: integer, format: int64 }
        rol: { $ref: "#/components/schemas/RolArchivo" }
        orden: { type: integer }
    ContenidoEntregaResponse:
      type: object
      description: Texto completo de la entrega, reconstruido por re-extracción (sin costo de LLM).
      properties:
        entregaId: { type: string, format: uuid }
        archivos:
          type: array
          items: { $ref: "#/components/schemas/ArchivoContenidoResponse" }
    ArchivoContenidoResponse:
      type: object
      properties:
        archivoId: { type: string, format: uuid }
        nombreOriginal: { type: string }
        rol: { $ref: "#/components/schemas/RolArchivo" }
        secciones:
          type: array
          items: { $ref: "#/components/schemas/SeccionContenidoResponse" }
    SeccionContenidoResponse:
      type: object
      description: Sección de texto limpio (página, título, archivo o celda de código).
      properties:
        titulo: { type: [string, "null"] }
        lenguaje: { type: [string, "null"] }
        lineaInicio: { type: [integer, "null"] }
        lineaFin: { type: [integer, "null"] }
        texto: { type: string }
...